Introduces tdstkit-core to reduce code duplication
